backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
pyjwt>=2.0
jinja2>=3.0
python-multipart>=0.0.5
